Switch to TS and exports only ESmodule
