+++ .eslintignore
@@ -1,0 +1,3 @@
+dist
+prettier.config.js
+karma.config.js
