+++ examples/index.html
@@ -3,8 +3,6 @@
 <head>
   <meta charset="utf-8">
   <title>combobox-nav demo</title>
-  <!-- <script type="text/javascript" src="../dist/combobox-nav.umd.js"></script> -->
-  <script type="text/javascript" src="https://unpkg.com/@github/combobox-nav@latest/dist/combobox-nav.umd.js"></script>
   <style>[aria-selected="true"] { font-weight: bold; }</style>
 </head>
 <body>
@@ -22,17 +20,20 @@
     </ul>
   </form>
   <pre class="events"></pre>
-  <script type="text/javascript">
+
+  <script type="module">
+    // import {install, uninstall, stop, start} from '../dist/index.js'
+    import {install, uninstall, stop, start} from 'https://unpkg.com/@github/combobox-nav@latest/dist/index.js'
     const input = document.querySelector('input')
     const list = document.querySelector('ul')
-    comboboxNav.install(input, list)
+    install(input, list)
 
     function toggleList() {
       const hidden = input.value.length === 0
       if (hidden) {
-        comboboxNav.stop(input)
+        stop(input)
       } else {
-        comboboxNav.start(input)
+        start(input)
       }
       list.hidden = hidden
     }
@@ -40,7 +41,7 @@
     input.addEventListener('focus', toggleList)
     input.addEventListener('blur', () => {
       list.hidden = true
-      comboboxNav.stop(input)
+      stop(input)
     })
 
     const events = document.querySelector('.events')
