--- index.d.ts
@@ -1,4 +1,0 @@
-export function install(input: HTMLTextAreaElement | HTMLInputElement, list: HTMLElement): void;
-export function navigate(input: HTMLTextAreaElement | HTMLInputElement, list: HTMLElement, indexDiff?: -1 | 1): void;
-export function uninstall(input: HTMLTextAreaElement | HTMLInputElement, list: HTMLElement): void;
-export function clearSelection(input: HTMLTextAreaElement | HTMLInputElement, list: HTMLElement): void;
