+++ karma.config.js
@@ -1,7 +1,7 @@
 module.exports = function(config) {
   config.set({
     frameworks: ['mocha', 'chai'],
-    files: ['../dist/combobox-nav.umd.js', 'test.js'],
+    files: [{pattern: 'dist/index.js', type: 'module'}, {pattern: 'test/test.js', type: 'module'}],
     reporters: ['mocha'],
     port: 9876,
     colors: true,
