+++ package.json
@@ -2,40 +2,34 @@
   "name": "@github/combobox-nav",
   "description": "Attach combobox navigation behavior to an input.",
   "version": "1.0.2",
-  "main": "dist/combobox-nav.umd.js",
-  "module": "dist/combobox-nav.esm.js",
-  "types": "index.d.ts",
+  "main": "dist/index.js",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
   "license": "MIT",
   "repository": "github/combobox-nav",
   "files": [
-    "dist",
-    "index.d.ts"
+    "dist"
   ],
   "scripts": {
     "clean": "rm -rf dist",
     "lint": "github-lint",
     "prebuild": "npm run clean && npm run lint && mkdir dist",
-    "build": "rollup -c && cp src/combobox-nav.js.flow dist/combobox-nav.esm.js.flow && cp src/combobox-nav.js.flow dist/combobox-nav.umd.js.flow",
-    "test": "karma start test/karma.config.js",
+    "build": "tsc",
+    "test": "karma start karma.config.js",
     "pretest": "npm run build",
     "prepublishOnly": "npm run build",
     "postpublish": "npm publish --ignore-scripts --@github:registry='https://npm.pkg.github.com'"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.0",
-    "babel-preset-github": "^3.2.1",
     "chai": "^4.2.0",
     "eslint": "^6.6.0",
-    "eslint-plugin-github": "^3.2.1",
-    "flow-bin": "^0.111.1",
-    "karma": "^4.4.1",
+    "eslint-plugin-github": "^3.4.1",
+    "karma": "^5.0.2",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-mocha": "^1.3.0",
+    "karma-mocha": "^2.0.0",
     "karma-mocha-reporter": "^2.2.5",
     "mocha": "^6.2.2",
-    "rollup": "^1.26.3",
-    "rollup-plugin-babel": "^4.3.2",
-    "rollup-plugin-node-resolve": "^5.2.0"
+    "typescript": "^3.8.3"
   }
 }
