+++ prettier.config.js
@@ -1,2 +1,1 @@
-/* @flow */
 module.exports = require('eslint-plugin-github/prettier.config')
