--- src/combobox-nav.js.flow
@@ -1,6 +1,0 @@
-/* @flow strict */
-
-declare export function install(input: HTMLTextAreaElement | HTMLInputElement, list: HTMLElement): void;
-declare export function navigate(input: HTMLTextAreaElement | HTMLInputElement, list: HTMLElement, indexDiff?: -1 | 1): void;
-declare export function uninstall(input: HTMLTextAreaElement | HTMLInputElement, list: HTMLElement): void;
-declare export function clearSelection(input: HTMLTextAreaElement | HTMLInputElement, list: HTMLElement): void;
